// touristagency-front/generated/models/Room.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tourist agency
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Hotel } from './Hotel';
import {
    HotelFromJSON,
    HotelFromJSONTyped,
    HotelToJSON,
} from './Hotel';
import type { ServiceClass } from './ServiceClass';
import {
    ServiceClassFromJSON,
    ServiceClassFromJSONTyped,
    ServiceClassToJSON,
} from './ServiceClass';

/**
 * 
 * @export
 * @interface Room
 */
export interface Room {
    /**
     * 
     * @type {string}
     * @memberof Room
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Room
     */
    name: string;
    /**
     * 
     * @type {ServiceClass}
     * @memberof Room
     */
    serviceClass: ServiceClass;
    /**
     * 
     * @type {number}
     * @memberof Room
     */
    pricePerNight: number;
    /**
     * 
     * @type {Hotel}
     * @memberof Room
     */
    hotel: Hotel;
}

/**
 * Check if a given object implements the Room interface.
 */
export function instanceOfRoom(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "serviceClass" in value;
    isInstance = isInstance && "pricePerNight" in value;
    isInstance = isInstance && "hotel" in value;

    return isInstance;
}

export function RoomFromJSON(json: any): Room {
    return RoomFromJSONTyped(json, false);
}

export function RoomFromJSONTyped(json: any, ignoreDiscriminator: boolean): Room {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'serviceClass': ServiceClassFromJSON(json['serviceClass']),
        'pricePerNight': json['pricePerNight'],
        'hotel': HotelFromJSON(json['hotel']),
    };
}

export function RoomToJSON(value?: Room | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'serviceClass': ServiceClassToJSON(value.serviceClass),
        'pricePerNight': value.pricePerNight,
        'hotel': HotelToJSON(value.hotel),
    };
}