// touristagency-front/generated/models/RoomIn.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tourist agency
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { ServiceClass } from './ServiceClass';
import {
    ServiceClassFromJSON,
    ServiceClassFromJSONTyped,
    ServiceClassToJSON,
} from './ServiceClass';

/**
 * 
 * @export
 * @interface RoomIn
 */
export interface RoomIn {
    /**
     * 
     * @type {string}
     * @memberof RoomIn
     */
    name: string;
    /**
     * 
     * @type {ServiceClass}
     * @memberof RoomIn
     */
    serviceClass: ServiceClass;
    /**
     * 
     * @type {number}
     * @memberof RoomIn
     */
    pricePerNight: number;
    /**
     * 
     * @type {string}
     * @memberof RoomIn
     */
    hotelId: string;
}

/**
 * Check if a given object implements the RoomIn interface.
 */
export function instanceOfRoomIn(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "serviceClass" in value;
    isInstance = isInstance && "pricePerNight" in value;
    isInstance = isInstance && "hotelId" in value;

    return isInstance;
}

export function RoomInFromJSON(json: any): RoomIn {
    return RoomInFromJSONTyped(json, false);
}

export function RoomInFromJSONTyped(json: any, ignoreDiscriminator: boolean): RoomIn {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'serviceClass': ServiceClassFromJSON(json['serviceClass']),
        'pricePerNight': json['pricePerNight'],
        'hotelId': json['hotelId'],
    };
}

export function RoomInToJSON(value?: RoomIn | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'serviceClass': ServiceClassToJSON(value.serviceClass),
        'pricePerNight': value.pricePerNight,
        'hotelId': value.hotelId,
    };
}